pyproject.toml
[project]
name = "s13a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "comtypes>=1.4.11",
    "dotenv>=0.9.9",
    "google-genai>=1.16.1",
    "google-generativeai>=0.3.0",
    "mss>=10.0.0",
    "numpy>=2.2.6",
    "onnx>=1.18.0",
    "onnxruntime-gpu>=1.22.0",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pywin32>=310",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "wget>=3.2",
    "mcp>=0.1.0",
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    "pyyaml>=6.0.1",        # For YAML config files
    "json-repair>=0.1.0",
    "pydantic>=2.6.1"      # For data validati
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "windowcontrol"}
packages = ["windowManager"]

[tool.setuptools.package-data]
"windowManager" = ["*.json", "*.txt"]